requirements.txt
datetime
django
django-cryptography
django-sockets
matplotlib
pylatexenc
pycryptodome
qiskit
qiskit-aer